test: initial conversion of test to typescript
